requirements.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
beautifulsoup4==4.8.2
black==19.10b0
bs4==0.0.1
Click==7.0
Django==3.0.1
django-ckeditor==5.8.0
django-cleanup==4.0.0
django-crispy-forms==1.8.1
django-filter==2.2.0
django-js-asset==1.2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
django-stdimage==5.0.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pathspec==0.7.0
Pillow==6.2.1
progressbar2==3.47.0
pylint==2.4.4
python-utils==2.3.0
pytz==2019.3
regex==2019.12.20
reportlab==3.5.32
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.0
wrapt==1.11.2
